src/modules/ColorPanelList/reducer.ts
import { StateType, ActionTypes } from '@/modules/ColorPanelList/types';

/**
 * State
 */
export const initialState: StateType = {
  items: [],
};

/**
 * Reducer
 */
export default function reducer(
  state: StateType = initialState,
  action: ActionTypes
): StateType {
  switch (action.type) {
    case 'addColorPanel':
      return {
        ...state,
        items: [...state.items, action.payload],
      };
    case 'removeColorPanel':
      return {
        ...state,
        items: state.items.filter(item => {
          return item.id !== action.payload;
        }),
      };
    default:
      return state;
  }
}
